requirements.txt
streamlit
google-generativeai
scikit-learn
numpy